Move tests to permission file
